+++ src/pages/Collections.js
@@ -1,119 +1,114 @@
-import axios from 'axios';
-import { Button } from 'react-bootstrap'
-import { useEffect, useState } from 'react';
-import { web3Provider, connectWallet, onNetworkUpdate, toUnitAmount } from '../utils/constants.js';
-import { OpenSeaPort, Network } from 'opensea-js';
-import { OrderSide } from 'opensea-js/lib/types';
+import { useEffect, useState } from "react";
+import { web3Provider } from "../utils/constants.js";
 
-import CheckoutModal from './CheckoutModal'
-import Carousel from "../components/Carousel"
+import { OpenSeaPort, Network } from "opensea-js";
+import * as Sea from "../services/Sea";
 
-const Collections = () => {
-	// page content
-	const [collectionResponse, setCollectionResponse] = useState([]);
-	const [items, setItems] = useState([]);
-	const [accountAddress, setAccountAddress] = useState([]);
-	const [collectionName, setCollectionName] = useState([]);
-	const [collectionBannerImage, setCollectionBannerImage] = useState([]);
-	const [collectionFeatureImage, setCollectionFeatureImage] = useState([]);
-	const [collectionImage, setCollectionImage] = useState([]);
-	const [collectionDescription, setCollectionDescription] = useState([]);
-	const [selectedName, setSelectedName] = useState([]);
-  const [selectedPrice, setSelectedPrice] = useState([]);
-  const [selectedIndex, setSelectedIndex] = useState([]);
-  const [show, setShow] = useState(false);
+
+
+const Home = () => {
+  const [items, setItems] = useState([]);
+  const [accountAddress, setAccountAddress] = useState([]);
+  const [collectionName, setCollectionName] = useState([]);
+  const [collectionBannerImage, setCollectionBannerImage] = useState([]);
+  const [collectionFeatureImage, setCollectionFeatureImage] = useState([]);
+  const [collectionImage, setCollectionImage] = useState([]);
+  const [collectionDescription, setCollectionDescription] = useState([]);
 
   let seaport = null;
 
-	useEffect(async () => {
-	
-		onChangeAddress();
-	const collectionItems = ['carbureted', 'nycmoments','jersey-shore',  'pixelmeows'];
-	//'toabs','xarb-io','cryptopunk','projectnft','catnfttest5','youyan','ooplace-collections-v2',,'land-dao-2','rocktest-v3'];
-	for (let i = 0; i < collectionItems.length; i++) {
-		//console.log(collectionItems[i]);
-  		fetchList(collectionItems[i]);
-}
+  useEffect(() => {
+    onChangeAddress();
+    async function getList() {
+      await fetchList();
+      if (!accountAddress) {
+        //await connectWallet();
+      }
+    }
+    getList();
+  }, []);
 
-		//console.log(collectionResponse);
-		setCollectionResponse(collectionResponse);
-		
-	}, [])
+  const onChangeAddress = () => {
+    console.log("calling onChangeAddress");
+    seaport = new OpenSeaPort(web3Provider, {
+      networkName: Network.Rinkeby,
+    });
+    const web3 = seaport.web3;
 
-	const onChangeAddress = () => {
-		//console.log('calling onChangeAddress')
-		seaport = new OpenSeaPort(web3Provider, {
-			networkName: Network.Rinkeby
-		})
-		const web3 = seaport.web3
+    web3.eth.getAccounts((err, res) => {
+      if (!err) {
+        setAccountAddress(res[0]);
+      }
+    });
+  };
 
-		web3.eth.getAccounts((err, res) => {
-			if (!err) {
-				setAccountAddress(res[0]);
-			}
+  const fetchList = async () => {
+    try {
+	  const params = new URLSearchParams(window.location.search);
+console.log(params);
+      let data = await Sea.fetchCollection(params.get('collection'));
+      let items = data.assets;
+      setItems(items);
 
-		})
-			}
+      setCollectionBannerImage(items[0].collection.banner_image_url);
+      setCollectionFeatureImage(items[0].collection.featured_image);
+      setCollectionImage(items[0].collection.image_url);
+      setCollectionName(items[0].collection.name);
+      setCollectionDescription(items[0].collection.description);
+      console.log(data.assets);
+    } catch (error) {
+      console.log("error getting collection:" + error);
+      // display error to user here
+    }
+  };
 
-	const buildDetailUrl = (assetAddress, tokenId) => {
-		//console.log(" asset address " + assetAddress);
-		return '/detail?asset_contract_address=' + assetAddress + '&token_id=' + tokenId;
+  const buildDetailUrl = (assetAddress, tokenId) => {
+    //console.log(" asset address " + assetAddress);
+    return (
+      "/detail?asset_contract_address=" + assetAddress + "&token_id=" + tokenId
+    );
+  };
 
-	}
-	const buildCollectionUrl = (collection) => {
-		//console.log(" asset address " + assetAddress);
-		return '/home?collection=' + collection;
+  const listItems = items.map((item, index) => (
+    <div className="product-card">
+      <div className="product-image">
+        <a href={buildDetailUrl(item.asset_contract.address, item.token_id)}>
+          <img src={item.image_preview_url} alt="" />
+        </a>
+      </div>
+      <div className="product-info">
+        <h5>
+          <a href={buildDetailUrl(item.asset_contract.address, item.token_id)}>
+            {item.name}{" "}
+          </a>
+        </h5>
 
-	}
+        <p />
+      </div>
+    </div>
+  ));
 
-	const fetchList =  async collection => {
-		const url = "https://rinkeby-api.opensea.io/api/v1/assets?order_direction=desc&offset=0&limit=1&collection=" + collection
-		console.log('fetching data for '+ collection )
-		try {
-			let response = await axios.get(url);
-			let items = response.data.assets;
-			setItems(items);
-			collectionResponse.push(items);
-			setCollectionBannerImage(items[0].collection.banner_image_url);
-			setCollectionFeatureImage(items[0].collection.featured_image);
-			setCollectionImage(items[0].collection.image_url);
-			setCollectionName(items[0].collection.name);
-			setCollectionDescription(items[0].collection.description);
-			//console.log(response.data.assets);
-		} catch (err) {
-			//console.log('error fetching assets:' + err);
-		}
-	}
- 
-	const carouselItems =collectionResponse.map((subItems, sIndex) => 
-           
-			 
-          
-         
-
-	 subItems.map((item, index) =>
-<div class="banner-carousel">
- <a href={buildCollectionUrl(item.collection.slug)}>{item.collection.name}</a>
-<img class="banner_image" src={item.collection.banner_image_url} />
-
-</div>
-
-	));
-
-  
-	return (
-     <div>
-             <Carousel
-        show={3}
-        infiniteLoop
-        withIndicator
-      >
-{carouselItems}
-            </Carousel>
+  return (
+    <div class="collection">
+      <div class="header-wrapper">
+        <div class="header-all-columns">
+          <img src={collectionBannerImage} alt="" />
         </div>
+        <div class="header-all-columns">
+          <h3 class="post-title">{collectionName}</h3>
+        </div>
+        <div class="header-left">
+          <img class="icon" src={collectionImage} align="left" alt="" />
+        </div>
+        <div class="header-right">
+          <div class="text-with-line-breaks">{collectionDescription}</div>
+        </div>
+      </div>
+      <div>
+        <section class="products">{listItems}</section>
+      </div>
+    </div>
+  );
+};
 
-		
-	)
-}
-
-export default Collections
+export default Home;
